cps support and "unit" option for speed
This update replaces the “matchcps” and “stretchTo” parameters with a
“unit” parameter that modified what “speed” means to Dirt.  Options for
speed are “rate” (original behavior, the default), “sec”, and “cps”.
Only the first character is needed and is case-independent.

sound “padlong” |+| speed “5” |+| unit “s” stretches the sample to 5
seconds

sound “loops125” |+| unit “c” will use the (hidden) cps parameter to
automatically stretch the sample to fit in one cycle, whatever the
current cps is set to (since speed defaults to “1”).  Adding speed “2”
would speed the sample to fit 2 samples per second.