command line option for gain, gain comp for shape
I added a command line option “-g, —gain” to change the overall gain of
Dirt when starting it up.  This applies regardless of whether the dirty
compression is turned on or off.  It should default to the old values.
The argument is given in dB from peak.

I also included some gain compensation for the “shape” parameter, so
the distortion doesn’t just increase the loudness. The compensation is
based on “typical” sounds, so results may vary, and it breaks down when
the shape parameter is near 1.

Finally, I made a small fix to server.c to avoid message spam when Dirt
receives extra OSC params.